code reorg changes
